add cookie factory docs and tests
